+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.91.0](https://github.com/sinameraji/kimiflare/compare/v0.90.1...v0.91.0) (2026-07-08)
+
+
+### Features
+
+* **cloud:** restore KimiFlare Cloud adapter with K2.7 free tokens ([#586](https://github.com/sinameraji/kimiflare/issues/586)) ([634406d](https://github.com/sinameraji/kimiflare/commit/634406deac5347baff1cfb49405a9053b1327b7f))
+
+## [0.90.1](https://github.com/sinameraji/kimiflare/compare/v0.90.0...v0.90.1) (2026-07-04)
+
+
+### Bug Fixes
+
+* **ui:** show loading state while resuming session and auto-display summary ([#604](https://github.com/sinameraji/kimiflare/issues/604)) ([87bda6b](https://github.com/sinameraji/kimiflare/commit/87bda6b636a64096ef21cc975e69621a4d2f9a8f))
+
+## [0.90.0](https://github.com/sinameraji/kimiflare/compare/v0.89.1...v0.90.0) (2026-07-01)
+
+
+### Features
+
+* **ui:** show session context summary on /resume ([#601](https://github.com/sinameraji/kimiflare/issues/601)) ([0c56f8e](https://github.com/sinameraji/kimiflare/commit/0c56f8e97e839fa1e334f9ecab9d8ae8db099e82))
+
 ## [0.89.1](https://github.com/sinameraji/kimiflare/compare/v0.89.0...v0.89.1) (2026-06-30)